better types
